Don't attempt to run installers every time, it is annoying.

+++ installers/vim.sh
@@ -7,26 +7,22 @@ $CP "$DIR/../configuration/vimrc" ~/.vimrc
 
 "$DIR/vim/neobundle.sh"
 
-# https://github.com/jelera/vim-javascript-syntax
-printf "\nInstalling javascript syntax\n"
-mkdir -p ~/.vim/syntax
-curl -0 https://raw.github.com/jelera/vim-javascript-syntax/master/syntax/javascript.vim > ~/.vim/syntax/javascript.vim
+if [ ! -f ~/.vim/syntax/javascript.vim ]; then
+    # https://github.com/jelera/vim-javascript-syntax
+    printf "\nInstalling javascript syntax\n"
+    mkdir -p ~/.vim/syntax
+    curl -0 https://raw.github.com/jelera/vim-javascript-syntax/master/syntax/javascript.vim > ~/.vim/syntax/javascript.vim
 
-mkdir -p ~/.vim/ftdetect
-curl -0 https://raw.github.com/jelera/vim-javascript-syntax/master/ftdetect/javascript.vim > ~/.vim/ftdetect/javascript.vim
-
-printf "\nInstalling jsctags\n"
-JSCTAGS_DIR=/usr/local/jsctags
-if [[ ! -d $JSCTAGS_DIR ]]; then
-    cd /usr/local
-    sudo git clone --recursive https://github.com/faceleg/doctorjs.git jsctags
-    cd jsctags
-    sudo make install
+    mkdir -p ~/.vim/ftdetect
+    curl -0 https://raw.github.com/jelera/vim-javascript-syntax/master/ftdetect/javascript.vim > ~/.vim/ftdetect/javascript.vim
 fi
 
-#cd ~/.vim/bundle/YouCompleteMe
-#./install.sh --clang-completer
-
 # CSSLint for Syntastic
-sudo npm install -g csslint
+command_exists () {
+    type "$1" &> /dev/null ;
+}
+
+if ! command_exists csslint; then
+    sudo npm install -g csslint
+fi
